Move Help & Specialisation to command package
