+++ src/main/java/seedu/duke/command/Help.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.command;
 
 public class Help {
     public static void displayHelpFile() {
