+++ src/main/java/seedu/duke/command/Specialisation.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.command;
 
 public class Specialisation {
